TODO: Clean styles
